+++ webextension/popup.js
@@ -503,6 +503,7 @@ function handleCheckResult(response, tabs, callback) {
         renderStatus(Tools.escapeHtml(DOMPurify.sanitize(response.message)));
         return;
     }
+    track(response.url);
     getCheckResult(response.markupList, function(resultText) {
         renderMatchesToHtml(resultText, response, tabs, callback);
     }, function(errorMessage, errorMessageCode) {
@@ -603,6 +604,25 @@ function doCheck(tabs) {
             lastCheck: new Date().getTime()
         }, function() {});
     });
+}
+
+function track(pageUrl) {
+    // for now, only track in Chrome until we have updated the privacy policy for Firefox:
+    if (Tools.isChrome()) {
+        let shortenedUrl = pageUrl.replace(/\?.*/, "<replaced>");  // for privacy reasons, don't log URL parameters
+        let url = encodeURIComponent(shortenedUrl);
+        let trackingUrl = "https://openthesaurus.stats.mysnip-hosting.de/piwik.php?idsite=12&rec=1&url=" +
+            url + "&action_name=check_text&rand=" + Date.now() + "&apiv=1";
+        let trackReq = new XMLHttpRequest();
+        trackReq.open('POST', trackingUrl);
+        trackReq.onerror = function() {
+            console.log("LT add-on tracking failed");
+        };
+        trackReq.ontimeout = function() {
+            console.log("LT add-on tracking failed with timeout");
+        };
+        trackReq.send();
+    }
 }
 
 document.addEventListener('DOMContentLoaded', function() {